idle_timeout does not appear to do what it says on the tin.
Based on the name, one would think that it controls the time after
which keystone would give up trying to talk to the databases, and 30
seconds seems to be a reasonable timeout for that.  However, that does
not appear to be the case based on the keystone code we are using:

    102     def get_engine(self):
    103         """Return a SQLAlchemy engine."""
    104         connection_dict = sqlalchemy.engine.url.make_url(CONF.sql.connection)
    105
    106         engine_args = {'pool_recycle': CONF.sql.idle_timeout,
    107                        'echo': False,
    108                        'convert_unicode': True
    109                        }

Instead, it seems that idle_timeout from the conf file is being
translated into the pool_recycle parameter to the sqlalhemy
connection, which is the time that sqlalchemy will forcibly close
connections.  Since we are not a terribly busy environment, I can
easily see where Keystone does not actaully talk to mysql for more
than 30 seconds, after which point the connections are forcibly closed
and we start getting SQL errors from Keystone.

For now, just comment out idle_timeout, which will wind up having
sqlalchemy use the defaults and never forcilbly close connections.