Update user passwords in keystone_register provider
When trying to create a user, if the user already exists, we were not
doing anything so far.

However, if the password is different, we need to udpate it. To avoid
changing the password with no reason, we try to get a token with the
password, and if it doesn't work, we update the password.

https://bugzilla.novell.com/show_bug.cgi?id=825622