Always enable cert_required when SSL is enabled.
If I understand
https://www.openssl.org/docs/ssl/SSL_CTX_set_verify.html#NOTES "The
certificate returned (if any) is checked." client certificates are
always only checked if the client actually submits a certificate. This
is also matched in the following articles:

http://www.linuxjournal.com/node/5487/print
http://etutorials.org/Programming/secure+programming/Chapter+10.+Public+Key+Infrastructure/10.7+Verifying+an+SSL+Peer+s+Certificate/

Thus it's actually preferred to set this to true. For comparison, nova,
cinder and glance always set ssl.CERT_REQUIRED (the effect of
cert_required=true) when a CA cert file is configured. IMO this is the
correct way to do it.