install .openrc in the keystone::server recipe
The .openrc file contains useful values for interacting with OpenStack
via the command line such as the keystone service's location and
credentials.

This is being moved from the nova cookbook since there are no nova
specific entries in the file and keystone is higher up in the
barclamp/cookbook dependency chain.