Added workaround for version requirement of keystonemiddleware
When using the v3 API  Keystonemiddleware (as in Juno) requires the auth_version
parameter to be set to "v3.0" instead of "v3" as in all the other places. This
adds a workaround to the KeystoneHelper that can be used by all the other
services' cookbooks that need configure keystonemiddleware.

(cherry picked from commit 73eb06ca535d70dfd0acf774d24695d00adbfb3b)
Signed-off-by: Ralf Haferkamp <rhafer@suse.de>